feat: fit new PR requirements
